Refactor spark interaction, move it to own component
[#169265877]

(it'll get much bigger soon)
