+++ src/components/view-3d/helpers.ts
@@ -1,0 +1,6 @@
+import { SimulationModel } from "../../models/simulation";
+
+export const PLANE_WIDTH = 1;
+
+export const planeHeight = (simulation: SimulationModel) =>
+  simulation.config.modelHeight * PLANE_WIDTH / simulation.config.modelWidth;
