+++ src/components/view-3d/spark-interaction.scss
